test: move generate_image tests to test_generate.py; doc timeout scope
Address CodeRabbit review on PR #27:
- Note in generate_image docstring that timeout_seconds only bounds the
  wait=True streaming path (wait=False uses a fixed short timeout).
- Relocate the four generate_image tests out of test_wrapper.py into a
  dedicated tests/test_generate.py, per the one-tool-group-per-file
  guideline. The shared streaming fakes (_FakeProc/_RecordingCtx/
  patched_stream/_OK_STREAM) move to tests/conftest.py so both files share
  them without duplication.
